requirements.txt
flask==2.3.3
bcrypt==4.0.1
mysql-connector-python==8.2.0
